### 2024-11-13 #7378 @rkurchin
👋 @larsenkg and/or @caldarolamartin, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
