### 2024-11-20 #7378 @uellue
Hi @rkurchin, hi @SnirMeiri, I don't have access to Matlab. Does the software work in Octave?  If yes, I could review this.
