### 2024-11-21 #7378 @SnirMeiri
Hi @uellue, thanks for your answer. I have no experience with Octave but since the application itself is written as a MATLAB application (.mlapp file) I am not sure it will be able to run on Octave. I have tried to run it on the Octave online compiler and it did not compile. Should the desktop version of Octave be able to run the file although the online version is not able to do so?
