### 2024-11-21 #7378 @uellue
@SnirMeiri with neither Matlab nor Octave I have much experience. I tried installing the Matlab runtime and running the mlapp file, but couldn't figure out how. To confirm, running it in the App Designer as described in the documentation is the only way? IMO it is unfortunate when an open source application depends so strongly on a runtime that is not easily accessible for everyone.

Long story short, I won't be able to review this submission because I currently can't run the software. It would be much better in general if the application runs stand-alone without purchasing a Matlab license and if the documentation explains how to do that.
