### 2024-11-21 #7378 @SnirMeiri
@uellue Thanks for your suggestions. Unfortunately as for now the application only runs through MATLAB. It was first written for internal use and only after our experimental collaborators have mentioned they believe it will be useful for other research groups  that we have decided to adjust it for more general use and make it public. I will try to see if I can make it a stand-alone application.
