### 2024-11-21 #7378 @fgjorup
> Hi @fgjorup. Do you have the Image Processing toolbox and the Parallel Computation toolbox installed? These are the only toolboxes needed which are not defaults in MATLAB. It is mentioned in the installation section in the readme file in github but not in the paper.md file. Should paper file have an installation section? If you do have these toolboxes installed and the functions still don't work I will try and see what might have happened but I have tried running it through different computers and MATLAB versions and it worked.

I just tried installing the two toolboxes, unfortunately my university has a very strict data policy,  so I cannot install software through external programs (matlab). Had it been python it would have been another story. I am afraid I won't be able to serve as reviewer.

-Frederik
