### 2025-01-07 #7378 @SnirMeiri
Hi @rkurchin,
I’ve gone through the pool of reviewers for MATLAB-based software in JOSS over the past few years and identified a few individuals who might be relevant: @jonbmartin, @tytell, @littlelazy6, and @ksyang2013.
I hope this list is helpful. If not, I’d be happy to look further for additional potential reviewers.
Thanks
