### 2025-01-08 #7378 @jonbmartin
Hi Kurchin,

I’m sorry, I am not available at the moment- I’ll be free to review in ~ a
month if you still need someone.

Jonathan

On Tue, Jan 7, 2025 at 12:23 PM Rachel Kurchin ***@***.***>
wrote:

> 👋 @jonbmartin <https://github.com/jonbmartin>, would you be willing to
> review this submission for JOSS? We carry out our checklist-driven reviews
> here in GitHub issues and follow these guidelines:
> https://joss.readthedocs.io/en/latest/review_criteria.html
>
