### 2025-02-05 #7748 @editorialbot
[REVIEW]: LCOI: Local crystalline orientation identifier using mask fitting implemented in MATLAB
**Submitting author:** @SnirMeiri (<a href="http://orcid.org/0000-0003-3133-8269">Snir Meiri</a>)
**Repository:** https://github.com/SnirMeiri/LCOI_local_crystalline_orientation_identifier
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
