### 2025-03-23 SnirMeiri/LCOI_local_crystalline_orientation_identifier#1 @aazocar
[Review] Reproducibility Issues in Directionality Detection
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

While testing the reproducibility of the examples, I found small differences in the output across the three example cases. The results after the directionality detection part of the examples walk-trough match overall but are not identical, after following the indicated input values for the `particle identification` process. Here are the three exported images for reference. I would like to clarify whether non-deterministic behavior is expected and document it.

![Image](https://github.com/user-attachments/assets/2fa9bceb-00df-4965-83f5-37956154e405)

![Image](https://github.com/user-attachments/assets/f86bfd7b-f534-47b7-865a-3b3ba0865af9)

![Image](https://github.com/user-attachments/assets/7241b061-eb3a-4a45-bfb8-3d51e3ebd968)
