### 2025-03-23 SnirMeiri/LCOI_local_crystalline_orientation_identifier#2 @aazocar
[Review] Improve Functionality Documentation with Detailed Descriptions
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

The current documentation provides a method description and a user guide covering the overall functionality of the software. However, there is no detailed breakdown of each key-value pair within the application, including expected inputs and outputs and functionality description. Without this, it was difficult to go beyond the examples provided and tune for different use cases.

It would be beneficial to add detailed documentation that allows users to find the information they need. Instead of keeping this documentation within README.md, having a dedicated file for it would also be an improvement.
