### 2025-03-23 SnirMeiri/LCOI_local_crystalline_orientation_identifier#3 @aazocar
[Review] Expand Example Usage
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

The example walkthrough for hexagonal, rectangular, and triangular cases is helpful, but it only covers particle detection in the `Main` tab and directionality detection in the `Mask Creation & Fitting` tab. It would be beneficial to extend these examples to include other aspects of direction analysis and strain calculations. Additionally, more descriptive documentation of the examples would improve clarity and usability.
