### 2025-03-23 SnirMeiri/LCOI_local_crystalline_orientation_identifier#4 @aazocar
[Review] Add Community Guidelines for Contributions, Issues, and Support
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

There are no clear guidelines for the community on how to contribute or report issues. Also adding some contact points for support would be good.
