### 2025-03-23 SnirMeiri/LCOI_local_crystalline_orientation_identifier#1 @SnirMeiri
Thanks for your comment.
Both the particle detection and directionality detection are deterministic. However, the result can depend on the particles that the user selects for the creation of the mask to be fitted, especially if the chosen unit-cell is strained/deformed. Screening the resulting fits and then optimizing the used mask by taking the average of the screened fits (using the optimize mask button) should reduce the effect of the initial choice. I will address this point in the reviewed documentation.
