### 2025-03-23 SnirMeiri/LCOI_local_crystalline_orientation_identifier#2 @SnirMeiri
I will work on improving the documentation and elaborate on the different values such that the user can tune the choice for different use cases.
