### 2025-03-23 SnirMeiri/LCOI_local_crystalline_orientation_identifier#5 @aazocar
[Review] Implementation of testing
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

I'm not completely sure about automated testing in MATLAB, however, it is my understanding it should be possible to have unit tests that could help catch potential issues before runtime.

https://www.mathworks.com/help/matlab/matlab_prog/ways-to-write-unit-tests.html

While testing the examples, I encountered a case where the selected mask had a point outside the image. Despite this, the `Identify Orientations` process still ran, resulting in an error. Implementing unit tests could help detect such issues early and improve the robustness of the software.
