### 2025-03-23 SnirMeiri/LCOI_local_crystalline_orientation_identifier#6 @aazocar
[Review] Strengthen State-of-the-Art Comparison
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

> State of the field: Do the authors describe how this software compares to other commonly-used packages?

Overall, the structure and description in the paper are well-written. However, the section discussing the state-of-the-art could be strengthened. While bond-orientational order parameters are mentioned, there is no discussion of specific software implementations. Are there software that have similar functionalities although with different methods? What are researchers using for this sort of analysis? Are there commercial software being used?
