### 2025-03-23 SnirMeiri/LCOI_local_crystalline_orientation_identifier#7 @aazocar
[Review] Missing Reference for MATLAB in Paper
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

> References: Is the list of references complete, and is everything cited appropriately that should be cited (e.g., papers, datasets, software)? Do references in the text use the proper [citation syntax](https://pandoc.org/MANUAL.html#extension-citations)?

The reference for MATLAB is missing. See here for more details:

https://de.mathworks.com/matlabcentral/answers/414438-how-do-i-cite-matlab-in-a-bibliography-or-a-published-journal
