### 2025-03-26 #7748 @jonbmartin
Hi @rkurchin and all I believe there was a mistake - this may have gotten lost in the other editorial thread but as I said previously I am not able to review this article, I think I was added anyways. I sincerely apologize, but I am already currently reviewing another JOSS article and I do not have bandwidth to review multiple articles from the same journal simultaneously.  I'm sorry if I contributed to miscommunication.

> I’m sorry, I am not available at the moment- I’ll be free to review in ~ a
> month if you still need someone.
>
> Jonathan
