### 2025-04-17 #7748 @rkurchin
👋 ziyiyin97 and/or archermarx, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
