### 2025-04-18 #7748 @rkurchin
👋 @raghurama123 and/or @ThFriedrich, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
