### 2025-04-19 #7748 @ThFriedrich
Hi @rkurchin
Yes, this seems interesting and within my expertise. I’ll review it shortly.
