### 2025-05-04 SnirMeiri/LCOI_local_crystalline_orientation_identifier#3 @SnirMeiri
I added explanations for the choice of parameters in the different examples and highlighted cases that require special attention (such as particles overlaps). In addition, I added a link to the user guide where the results of strain analysis and tessellation of the hexagonal unit-cell (using the same image) are demonstrated.
