### 2025-05-04 SnirMeiri/LCOI_local_crystalline_orientation_identifier#5 @SnirMeiri
I have added tests for the functions in the software in a folder named tests within the App and functions folder. All tests can be executed (within MATLAB) using the runAllTests.m file. As for the user interface, I have limited all click-based choice of locations to be accepted only within the relevant UI-axis to eliminate issues such as the one you have reported. In case where too many wrong attempts of point choice were made the callback is eliminated (but can be restarted). In addition, more conditions over the relative order of the actions of the user were added in order to ensure that all the required prior steps of the user were done before executing callbacks.
