### 2025-05-04 SnirMeiri/LCOI_local_crystalline_orientation_identifier#6 @SnirMeiri
I have expanded about used software and how these implementations are different from our approach. It can be found in the statement-of-need section.
