### 2025-05-14 SnirMeiri/LCOI_local_crystalline_orientation_identifier#8 @ThFriedrich
[Review] Improve User Guide Style
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

## User Guide Style
The User guide.md could be a bit refactored to help the clarity of  the document. Suggestions would be to:
 1. Use subheaders in accordance with the tabs in the GUI for quick reference navigation
 2. consider putting background info into foladable Info-boxes like so:

| Current  | Suggested |
| - | - |
| The next step is locating the particles, using the panel marked in red in the main tab. It is done using circular Hough transform (imfindcircles) with user specified parameters as minimal and maximal radius of expected particles, measured in pixels, and sensitivity which is a value between 0 and 1. The user can choose whether bright or dark particles are expected and the used method for imfindcircles between PhaseCode and TwoStage using the toggle switches. In general PhaseCode is the faster method and is more robust to noise but TwoStage is generally more sensitive and detects more circles at a given sensitivity value. In addition, thresholding of the edges can be done automatically, using manually specified threshold between 0 and 1, where smaller number corresponds to weaker edges, or using a scanning scheme in a specified range and equally spaced steps. Regarding the last option, the software can choose parameters automatically by reducing the radii range and changing the sensitivity and the manual thresholding value (depending on the threshold scroll-down value) until changing them does not result in higher number of identified particles. Note however that the search can be slow and the search time depends on the width of the user specified minimal and maximal particle radius and on the number of particles. Thus, narrowing the range manually and/or running first on a small cropped region is suggested. In addition, there is a distance measuring tool (1) in order to estimate the size of the particles. In order to increase the number of found particles by using the expected spatial information (i.e. the expected overlap of particles), the general approach in this scheme is locating particles by the Hough transformation with high sensitivity and then applying filtration. Since large particles overlaps are not expected, the found particles are screened according to a minimal distance specified by the user, where the better fitted particle is kept according to the resulting metric from the Hough transform. Two additional filtration steps are applied in order to minimize detection of the gaps between particles in case they are similar in size to the particles themselves. The first applies a smaller circular filter around the detected particles centers and penalizes high variance of pixels values and the second sums all pixels values using a circular disk with a radius of the mean identified radii. The first filtration step is used to minimize detection of particles' centers near the edges of the particles and the second scores the fit by integrating the pixels' values within the detected circle. Due to brevity of the GUI these values can be changed in the code. The parameters of the filtration are: the fraction of disk to be considered compared to the mean radius of the detected particles, the factor of the threshold metric to be kept (between 0 and 1), the power of the standard deviation in the score of the filter(factStd), the percentile of the metric which is set as a threshold (between 0 and 100), the power of the minimal pixel value in the score (fact1) and the power of the minimal pixel value in the score (fact2). The formula of the score reads: $modified \quad score=\frac{(minimal \quad pixel \quad value)^{fact1}}{(maximal \quad pixel \quad value)^{fact2} \cdot std^{factStd}}.$ The score is then normalized by the radius of the detected particle squared and only scores above the product of the chosen percentile of the score and the chosen factor are kept. The default values for the filtration steps are (0.5,0.15,4,90,0,0) and (1,0.33,0,90,0,0), respectively. Notice that normalizing by the maximal/minimal pixel value may skew the results and give sub-optimal centering (but are kept here as an option for the flexibility of the filter). Last, the obtained locations of the particles’ centers may be improved using center optimization of a circular disk filter where at each step the located center can move by up to 1/8 of the mean radius in the x and y directions. Notice that large particles overlaps may cause artifacts in the resulting locations.  In addition, the user can upload particles identification data from a different software, saved to a mat file. In order to do so, the particles centers in a (x,y) order should by saved in a N-by-2 array named centers, and the particles’ radii and identification metric should by saved as N-by-1 arrays named radii and particlesMetric, respectively. At any point, the user can show the resulting layers on the right panel using the checkboxes (2). Particles can be manually deleted in the Adjust contrast & delete particles tab. | The next step is locating the particles, using the panel marked in red in the main tab. It is done using circular Hough transform (imfindcircles) with user specified parameters as minimal and maximal radius of expected particles, measured in pixels, and sensitivity which is a value between 0 and 1. The user can choose whether bright or dark particles are expected and the used method for imfindcircles between PhaseCode and TwoStage using the toggle switches. In general PhaseCode is the faster method and is more robust to noise but TwoStage is generally more sensitive and detects more circles at a given sensitivity value. In addition, thresholding of the edges can be done automatically, using manually specified threshold between 0 and 1, where smaller number corresponds to weaker edges, or using a scanning scheme in a specified range and equally spaced steps. Regarding the last option, the software can choose parameters automatically by reducing the radii range and changing the sensitivity and the manual thresholding value (depending on the threshold scroll-down value) until changing them does not result in higher number of identified particles. Note however that the search can be slow and the search time depends on the width of the user specified minimal and maximal particle radius and on the number of particles. Thus, narrowing the range manually and/or running first on a small cropped region is suggested. In addition, there is a distance measuring tool (1) in order to estimate the size of the particles. <br><br> <details><summary>In depth info</summary>In order to increase the number of found particles by using the expected spatial information (i.e. the expected overlap of particles), the general approach in this scheme is locating particles by the Hough transformation with high sensitivity and then applying filtration. Since large particles overlaps are not expected, the found particles are screened according to a minimal distance specified by the user, where the better fitted particle is kept according to the resulting metric from the Hough transform. Two additional filtration steps are applied in order to minimize detection of the gaps between particles in case they are similar in size to the particles themselves. The first applies a smaller circular filter around the detected particles centers and penalizes high variance of pixels values and the second sums all pixels values using a circular disk with a radius of the mean identified radii. The first filtration step is used to minimize detection of particles' centers near the edges of the particles and the second scores the fit by integrating the pixels' values within the detected circle. Due to brevity of the GUI these values can be changed in the code. The parameters of the filtration are: the fraction of disk to be considered compared to the mean radius of the detected particles, the factor of the threshold metric to be kept (between 0 and 1), the power of the standard deviation in the score of the filter(factStd), the percentile of the metric which is set as a threshold (between 0 and 100), the power of the minimal pixel value in the score (fact1) and the power of the minimal pixel value in the score (fact2). The formula of the score reads: $modified \quad score=\frac{(minimal \quad pixel \quad value)^{fact1}}{(maximal \quad pixel \quad value)^{fact2} \cdot std^{factStd}}.$ The score is then normalized by the radius of the detected particle squared and only scores above the product of the chosen percentile of the score and the chosen factor are kept. The default values for the filtration steps are (0.5,0.15,4,90,0,0) and (1,0.33,0,90,0,0), respectively. Notice that normalizing by the maximal/minimal pixel value may skew the results and give sub-optimal centering (but are kept here as an option for the flexibility of the filter).</details><br> Last, the obtained locations of the particles’ centers may be improved using center optimization of a circular disk filter where at each step the located center can move by up to 1/8 of the mean radius in the x and y directions. Notice that large particles overlaps may cause artifacts in the resulting locations.  In addition, the user can upload particles identification data from a different software, saved to a mat file. In order to do so, the particles centers in a (x,y) order should by saved in a N-by-2 array named centers, and the particles’ radii and identification metric should by saved as N-by-1 arrays named radii and particlesMetric, respectively. At any point, the user can show the resulting layers on the right panel using the checkboxes (2). Particles can be manually deleted in the Adjust contrast & delete particles tab. |
