### 2025-05-14 SnirMeiri/LCOI_local_crystalline_orientation_identifier#9 @ThFriedrich
[Review] Check for installed toolboxes
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

Trying to use parallel computation when you don't have the toolbox installed leaves the user waiting, with the status indicator showing "busy". Perhaps one can check wether the toolbox is available and catch that error with something like:
```Octave
% inspired by: https://de.mathworks.com/matlabcentral/answers/419823-how-to-catch-warnings
% Infos on toolbox names for the ver()-command: https://de.mathworks.com/matlabcentral/answers/377731-how-do-features-from-license-correspond-to-names-from-ver#answer_430148
function b_check = check_for_toolbox(toolbox)
    if nargin == 0
        toolbox = 'parallel';
    end
    % reset the lastwarn message and id
    lastwarn('', '');

    % call ver, which return the version of the toolbox or a warning when not installed
    ver(toolbox)

    % see if a warning was raised
    if(isempty(lastwarn()))
        b_check = 1;
    else
        b_check = 0;
    end
end
```
