### 2025-05-14 SnirMeiri/LCOI_local_crystalline_orientation_identifier#10 @ThFriedrich
[Review] Installation instructions
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

The instructions given in the paper work in principle, but the AppDesigner isn't really meant to be used as a deployment platform. I would recommend to:
 1. Add an installable package, which can simply be created with the AppDesigner.
 2. Add the option to run the App without installation from the already available exported class object (i.e. just running ```LCOI_CrystalAnalysis_exported``` in the Matlab command prompt)
