### 2025-05-14 SnirMeiri/LCOI_local_crystalline_orientation_identifier#11 @ThFriedrich
[Review] Add CONTRIBUTING file
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

There are common approaches in the OS community to add guidelines for contributing, which I suggest to follow. Perhaps take a look at the documentation here:  [Setting guidelines for repository contributors](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)
