### 2025-05-14 #7748 @ThFriedrich
I've been working on the Review. First I'd like to express my appreciation for the contribution, which I am sure will be helpful to the microscopy community. The effort that was put into creating not only the algorithm, but also wrapping it into an easy-to-use framework and creating the GUI application is substantial. Thank you for the effort @SnirMeiri . I opened a few Issues over at the [original repository](https://github.com/SnirMeiri/LCOI_local_crystalline_orientation_identifier), but I'd like to add a few notes to the authors here too, which may be less relevant for this particular publication, but may be useful nevertheless.

## Documentation
There is no classical API documentation, which I guess is fine since there is no API by design. Using docstrings one can get some basic documentation on the class with:
```Ocatve
doc LCOI_CrystalAnalysis_exported
```
which is not great because its all intermixed with the GUI functions. I am not sure how the development process for Matlab-GUI Apps is, but a paradigm where almost all the functionality of such a relatively complex program is cramped into a single class seems not ideal. It is quite cumbersome to seperate base functionality and GUI-related functions (Button clicks, callbacks, etc). It also makes it difficult for other developers to join the project, because there is no apparent structure from the filetree. You could work with [namespaces and classes](https://de.mathworks.com/help/matlab/matlab_oop/namespaces.html) and things like that in matlab too to get a more object-oriented kind of structure. If you do that (in combination with your docstrings), you get the API and API documentation for free. ;-)
That being said, the code looks clean and functions are documented in the code, which is nice. Given that its a GUI application and no API I find this acceptable overall. Refactoring the whole thing would be quite some work which, depending wether you want to build much further on this code or not, you could consider in the future. Otherwise its just advice for future projects. :-)

## Broader scope
I wonder if you tried/considered using this approach for atomic-resolution TEM and STEM data? I used to work on the [StatSTEM](https://github.com/quantitativeTEM/StatSTEM) project for example, which I think could benefit from such an approach for indexing atomic columns. (Maybe its worth for @annickdebacker or @pho-Toms to have a look at this)

## Quality of writing
In the paper, but especially in the user guide, the Quality of writing should be improved somewhat I think. While it is not bad, there are a few minor mistakes and lengthy, nested sentences. Maybe a native speaker or some AI could help improve the flow a bit?
