### 2025-05-16 SnirMeiri/LCOI_local_crystalline_orientation_identifier#13 @ThFriedrich
[Review] Colormap granularity
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

During testing I noticed, that changing the granularity of the colormap for orientation plotting leads to overlayed legends. I assume the figure isn't getting cleared, which I assume is not the inteded beahaviour.

Here is an example where I did 20, 8 and 21 colorsteps:
![Image](https://github.com/user-attachments/assets/d4b17cb7-6307-4306-a5e5-3ebe213af725)

Further, in the user guide it states: "The value of the number of colors in the colormap, chosen by the user, changes the granularity of the colormap and does not affect the colorcoding itself. ". However, it seems one has to Re-Identify orientations to actually update the colors.
