### 2025-05-16 SnirMeiri/LCOI_local_crystalline_orientation_identifier#14 @ThFriedrich
[Review] Plots not restored when opening project
This is part of the review at https://github.com/openjournals/joss-reviews/issues/7748

When opening a previously saved project, the last plots are not automatically restored, which would be a nice feature, so one can see right away what the progress or last status of the project was. This is just an enhancement idea, not really an issue. The data is all there, one just has to replot manually.
