+++ paper.md
@@ -23,12 +23,15 @@ Detection and analysis of the orientation of local crystalline order arises in d
 Several software packages analyze local crystalline order based solely on real-space imaging data. One widely used approach is the bond-orientational order parameter (BOOP), originally introduced by Steinhardt et al. [@Steinhardt1983], and implemented in various tools [@Stukowski2010;@pyscal;@freud2020], which provide quantitative tools to classify local environments. These tools are commonly used in both simulations and experimental particle tracking contexts. However, BOOP-based methods have notable limitations, including their reliance on isotropic assumptions, high sensitivity to defects, and a lack of robustness to small spatial fluctuations [@Mickel:2013].
 Some commercial tools also support symmetry and orientation analysis, especially from diffraction data. For example, CrysTBox [@Klinger:vh5045] use Fourier transforms and lattice fitting methods, primarily working in the frequency domain. It offers limited orientation mapping directly from HRTEM images via geometric phase analysis and Fourier-based methods.
 Our approach departs from both diffraction-based and BOOP-based methods. We propose an approach based on optimally fitting a mask of specific shape to the detected particles. Within this approach, the resulting directionality is continuous and robust to local noise. The accuracy in mask orientation detection is primarily limited by the accuracy of the particles localization, rather than from the symmetry assumptions or discrete angular jumps typical of BOOP-based metrics.
+
 # Summary
 LCOI is designed to extract orientation information about local crystalline order from microscopy images such as transmission electron microscopy (TEM). Specifically, the software is able to identify triangular, rectangular and hexagonal order within planar cross-sections of colloidal crystals. The orientation identification is based on mask fitting to the identified particles. The software, implemented as a MATLAB [@MATLAB] application and available via [GitHub](https://github.com/SnirMeiri/LCOI_local_crystalline_orientation_identifier), provides tools for basic image processing, particle identification, mask creation and fitting, orientation analysis, and strain measurement in crystals. The following figure summarizes the general flow of the identification process:
 ![](User%20guide%20images/over1.png)
 After loading the image and optionally adjusting the contrast, the user can apply particle detection or upload pre-identified data. The user then creates a mask (hull) by manually selecting particles of the desired symmetry, which is subsequently symmetrized by the software. The options are triangular, rectangular and hexagonal, where the last two can either have 2-fold symmetry or 4-fold and 6-fold, respectively. Next, the local orientation of the optimal fits of the created mask to the identified particles is detected and color-coded. A more detailed scheme of the orientation detection is:
 ![](User%20guide%20images/over2.png)
 After particles identification and mask creation, the following step is particle pairs identification by considering all particles below a distance threshold set by considering the created mask. Then, a fitting mask is located as if each pair corresponds to an edge of the mask. These initial guesses are screened for the sum of the distances to the closest particles to each mask's vertex.  Next, for the screened masks’ location and orientation, the particle associated with each mask’s vertex is identified and the optimal location and orientation is computed analytically. For more details see the attached file: Mask location and rotation optimization. Finally, the mask can be refined using all identified fits by averaging the relevant distances, after which the optimal location and orientation are recomputed.
+
 # Acknowledgements
 We gratefully acknowledge Prof. Rafal Klajn and his group members for introducing us to the field and its challenges which resulted in this software. This work was supported by ISF grant No. 1444/21.
+
 # References
