### 2025-05-18 SnirMeiri/LCOI_local_crystalline_orientation_identifier#8 @SnirMeiri
I added sub-headers for simplifying the navigation in the user guide and put some of the extra details in foldable boxes to improve readability. In addition, I tried to simplify sentences/improve language in the file.
Thank you for these suggestions.
