### 2025-05-18 SnirMeiri/LCOI_local_crystalline_orientation_identifier#9 @SnirMeiri
I added a verification for the two toolboxes. The user is notified if any of them is missing when running the file and if the user tries to use parallel computation but the package is missing, serial computation is being used and the user is notified.
