### 2025-05-18 SnirMeiri/LCOI_local_crystalline_orientation_identifier#10 @SnirMeiri
An installable package was uploaded and the different options for running the app were updated in the readme file.
