### 2025-05-18 SnirMeiri/LCOI_local_crystalline_orientation_identifier#11 @SnirMeiri
I added a contributing file. I attempted to match a common approach for the guidelines.
