### 2025-05-18 SnirMeiri/LCOI_local_crystalline_orientation_identifier#13 @SnirMeiri
Thanks for finding this bug of the overlaid legends. It was fixed.
About the last remark maybe there is a misunderstanding. The granularity only affects the legend of the plot but not the colors in the plot itself.
