### 2025-05-18 SnirMeiri/LCOI_local_crystalline_orientation_identifier#14 @SnirMeiri
I added restoration of a few plots when opening a saved project. Since two or more unit-cell types can be identified in a single image some of the plots are only plotted if a single type was identified.
