### 2025-05-18 #7748 @SnirMeiri
Hi @ThFriedrich,
First I would like to thank you for the time you spent reviewing the app and paper. I have addressed all the issues you have raised so far (I commented on each separately) and I believe that it helped improving both the application and the paper/repository auxiliary files.
About your advice on the structure of the code, I will try to implement it in my next project. This app was not originally made to be published as open source code, but rather as in-house tool for analysis in a joint project with our experimental collaborators. It was just after that our collaborators mentioned that they believe it will be helpful for other groups, that we have decided to make it public. In hind-sight, some of it's structure would have been designed differently.
As mentioned, this was written to analyze data, and in our case it was used to analyze HAADF-STEM images. I believe it will be useful for other similar types of data, depending on the resolution of the data.
Regarding your last comment on the style of writing I have tried to improve the flow of the text by simplifying some of the longer and nested sentences and I have corrected some mistakes/typos (also using AI).
Thanks again for your input and time.
