### 2025-05-18 SnirMeiri/LCOI_local_crystalline_orientation_identifier#13 @ThFriedrich
Thanks for fixing this. Indeed, I wasn't quite sure about the second point. Is there a particular reason for changing the granularity of the colormap key? Is there good reasons for not making it continuous?
