### 2025-05-18 SnirMeiri/LCOI_local_crystalline_orientation_identifier#8 @ThFriedrich
This looks very nice now. Thank you! 😄
