+++ User guide.md
@@ -56,7 +56,7 @@ The mask is created using the second tab, Mask creation & fitting:
 
 <img src="https://github.com/SnirMeiri/crystalDirections/blob/main/User%20guide%20images/masks.png" width="1400">
 
-The user should first plot the image and detected particles (1). The user can zoom in on a domain (for large crystals) by picking two points using the mouse and click on zoom-in. After clicking on pick particles (2) in the tab of the desired unit-cell type, the particles are chosen by clicking on the image close to the desired particles. The expected order for choosing particles is explained in the instructions text-box. Then, the user should click on create mask (3). Next, the user can click on identify orientations (4) while using the numeric factors in the Screening factors panel. The user can turn parallel computation on and off using the switch, where non-parallel is faster for relatively small images. Last, the user can choose to optimize the fitting mask. The value of the number of colors in the colormap, chosen by the user, changes the granularity of the colormap and does not affect the colorcoding itself. The user can sort (5) the found masks using two parameters, multiplicative and fraction. 
+The user should first plot the image and detected particles (1). The user can zoom in on a domain (for large crystals) by picking two points using the mouse and click on zoom-in. After clicking on pick particles (2) in the tab of the desired unit-cell type, the particles are chosen by clicking on the image close to the desired particles. The expected order for choosing particles is explained in the instructions text-box. Then, the user should click on create mask (3). Next, the user can click on identify orientations (4) while using the numeric factors in the Screening factors panel. The user can turn parallel computation on and off using the switch, where non-parallel is faster for relatively small images. Last, the user can choose to optimize the fitting mask. The user can sort (5) the found masks using two parameters, multiplicative and fraction. 
 
 <details><summary>Sorting parameters</summary> All masks whose sum distance between their vertices and the closest particles’ centers is lesser or equal to $M \cdot(\frac{mean \quad radius}{F})^2$  are kept, where $M$ is the Multiplicative  factor and $F$ is the fraction factor. $M$ usually corresponds to the number of particles in the expected unit-cell and to whther the user wants to keep fits with fewer particles (corresponding to defects) and $F$ corresponds to the expected variance around the used fitting mask where larger values correspond to higher thresholding of the kept fits.</details>
   
