### 2025-05-19 SnirMeiri/LCOI_local_crystalline_orientation_identifier#13 @SnirMeiri
I have changed to a continuous color-wheel and updated the relevant figures in the paper, readme and user guide.
