### 2025-05-20 #7748 @aazocar
@rkurchin I've completed my review and verified all checklist items. All the raised issues were thoroughly addressed, and I particularly appreciate the efforts by @SnirMeiri in implementing more comprehensive testing and improving software robustness and user guide.

There remain minor editorial points concerning the paper: the lack of figure captions and the use of abbreviations without prior definition (e.g., line 22, "HRTEM"). Apart from these minor comments, my review is concluded.

Thanks to the authors for their efforts and making their software openly available to the community!
