+++ paper.md
@@ -25,12 +25,12 @@ Some commercial tools also support symmetry and orientation analysis, especially
 Our approach departs from both diffraction-based and BOOP-based methods. We propose an approach based on optimally fitting a mask of specific shape to the detected particles. Within this approach, the resulting directionality is continuous and robust to local noise. The accuracy in mask orientation detection is primarily limited by the accuracy of the particles localization, rather than from the symmetry assumptions or discrete angular jumps typical of BOOP-based metrics.
 
 # Summary
-LCOI is designed to extract orientation information about local crystalline order from microscopy images such as transmission electron microscopy (TEM). Specifically, the software is able to identify triangular, rectangular and hexagonal order within planar cross-sections of colloidal crystals. The orientation identification is based on mask fitting to the identified particles. The software, implemented as a MATLAB [@MATLAB] application and available via [GitHub](https://github.com/SnirMeiri/LCOI_local_crystalline_orientation_identifier), provides tools for basic image processing, particle identification, mask creation and fitting, orientation analysis, and strain measurement in crystals. The following figure summarizes the general flow of the identification process:
+LCOI is designed to extract orientation information about local crystalline order from microscopy images such as transmission electron microscopy (TEM). Specifically, the software is able to identify triangular, rectangular and hexagonal order within planar cross-sections of colloidal crystals. The orientation identification is based on mask fitting to the identified particles. The software, implemented as a MATLAB [@MATLAB] application and available via [GitHub](https://github.com/SnirMeiri/LCOI_local_crystalline_orientation_identifier), provides tools for basic image processing, particle identification, mask creation and fitting, orientation analysis, and strain measurement in crystals. 
 
-![Schematic representation of the orientation identification process. The result demonstrates the color-coded orientations of the identified 2-fold symmetric hexagonal crystalline order.](User%20guide%20images/over1.png)
+![Summary of the general work-flow of the orientation identification process, demonstrated using TEM data. The result shows the color-coded orientations of the identified 2-fold symmetric hexagonal crystalline order.](User%20guide%20images/over1.png)
 
 After loading the image and optionally adjusting the contrast, the user can apply particle detection or upload pre-identified data. The user then creates a mask (hull) by manually selecting particles of the desired symmetry, which is subsequently symmetrized by the software. The options are triangular, rectangular and hexagonal, where the last two can either have 2-fold symmetry or 4-fold and 6-fold, respectively. Next, the local orientation of the optimal fits of the created mask to the identified particles is detected and color-coded. A more detailed scheme of the orientation detection is:
-![](User%20guide%20images/over2.png)
+![Schematic representation of the process of optimal fitting of the masks.](User%20guide%20images/over2.png)
 After particles identification and mask creation, the following step is particle pairs identification by considering all particles below a distance threshold set by considering the created mask. Then, a fitting mask is located as if each pair corresponds to an edge of the mask. These initial guesses are screened for the sum of the distances to the closest particles to each mask's vertex.  Next, for the screened masks’ location and orientation, the particle associated with each mask’s vertex is identified and the optimal location and orientation is computed analytically. For more details see the attached file: Mask location and rotation optimization. Finally, the mask can be refined using all identified fits by averaging the relevant distances, after which the optimal location and orientation are recomputed.
 
 # Acknowledgements
