### 2025-05-20 #7748 @SnirMeiri
@aazocar
I added prior definition to HRTEM and captions to the figures.
Thank you very much for your time and input.
