### 2025-05-20 #7748 @rkurchin
Editorial notes...

### paper
- line 7: missing Oxford comma after "materials science"
- 8: same after "twinning"
- 27: need an apostrophe after "particles" or change to singular "particle"
- figure 1: also suggest changing upper right label to "particle identification" (likewise upper left in figure 2)
- figure 1 caption: "workflow" is usually one (unhyphenated) word
- figure 2: would change to "pair identification"
- 43: "particle identification"
- 43, later, suggest changing to "identification of pairs of particles" or just "pair identification"
- 49: add comma after "details" and also add more clarity about where to find the file (JOSS doesn't have "supplementary information" in the traditional way, so I would suggest uploading it to your repository or including it in your software documentation and directing readers there)

### references
- fix MATLAB citation so it doesn't render as if "Inc." is a surname
