+++ paper.md
@@ -19,21 +19,21 @@ date: 25 September 2024
 bibliography: paper.bib
 ---
 # Statement of need
-Detection and analysis of the orientation of local crystalline order arises in different fields such as chemistry, materials science and condensed matter physics. In particular, such analysis is central for studying crystal melting, jamming, crystal twinning and crystal strains, among others. Although the local orientations can be deduced from the diffraction data [@Busing:1967; @Zaefferer:2011; @Angel:2015], such analysis must take some spatial averages.
+Detection and analysis of the orientation of local crystalline order arises in different fields such as chemistry, materials science, and condensed matter physics. In particular, such analysis is central for studying crystal melting, jamming, crystal twinning, and crystal strains, among others. Although the local orientations can be deduced from the diffraction data [@Busing:1967; @Zaefferer:2011; @Angel:2015], such analysis must take some spatial averages.
 Several software packages analyze local crystalline order based solely on real-space imaging data. One widely used approach is the bond-orientational order parameter (BOOP), originally introduced by Steinhardt et al. [@Steinhardt1983], and implemented in various tools [@Stukowski2010;@pyscal;@freud2020], which provide quantitative tools to classify local environments. These tools are commonly used in both simulations and experimental particle tracking contexts. However, BOOP-based methods have notable limitations, including their reliance on isotropic assumptions, high sensitivity to defects, and a lack of robustness to small spatial fluctuations [@Mickel:2013].
 Some commercial tools also support symmetry and orientation analysis, especially from diffraction data. For example, CrysTBox [@Klinger:vh5045] use Fourier transforms and lattice fitting methods, primarily working in the frequency domain. It offers limited orientation mapping directly from high-resolution transmission electron microscopy images via geometric phase analysis and Fourier-based methods.
-Our approach departs from both diffraction-based and BOOP-based methods. We propose an approach based on optimally fitting a mask of specific shape to the detected particles. Within this approach, the resulting directionality is continuous and robust to local noise. The accuracy in mask orientation detection is primarily limited by the accuracy of the particles localization, rather than from the symmetry assumptions or discrete angular jumps typical of BOOP-based metrics.
+Our approach departs from both diffraction-based and BOOP-based methods. We propose an approach based on optimally fitting a mask of specific shape to the detected particles. Within this approach, the resulting directionality is continuous and robust to local noise. The accuracy in mask orientation detection is primarily limited by the accuracy of the particles' localization, rather than from the symmetry assumptions or discrete angular jumps typical of BOOP-based metrics.
 
 # Summary
 LCOI is designed to extract orientation information about local crystalline order from microscopy images such as transmission electron microscopy (TEM). Specifically, the software is able to identify triangular, rectangular and hexagonal order within planar cross-sections of colloidal crystals. The orientation identification is based on mask fitting to the identified particles. The software, implemented as a MATLAB [@MATLAB] application and available via [GitHub](https://github.com/SnirMeiri/LCOI_local_crystalline_orientation_identifier), provides tools for basic image processing, particle identification, mask creation and fitting, orientation analysis, and strain measurement in crystals. 
 
-![Summary of the general work-flow of the orientation identification process, demonstrated using TEM data. The result shows the color-coded orientations of the identified 2-fold symmetric hexagonal crystalline order.](User%20guide%20images/over1.png)
+![Summary of the general workflow of the orientation identification process, demonstrated using TEM data. The result shows the color-coded orientations of the identified 2-fold symmetric hexagonal crystalline order.](User%20guide%20images/over1.png)
 
[…]
