+++ paper.bib
@@ -47,7 +47,7 @@
 
 @software{MATLAB,
   year      = {2023},
-  author    = {The MathWorks Inc.},
+  author    = {{The MathWorks Inc.}},
   title     = {MATLAB version: 9.14.0 (R2023a)},
   publisher = {The MathWorks Inc.},
   address   = {Natick, Massachusetts, United States},
