+++ paper.bib
@@ -47,8 +47,8 @@
 
 @software{MATLAB,
   year      = {2023},
-  author    = {{The MathWorks Inc.}},
-  title     = {MATLAB version: 9.14.0 (R2023a)},
+  author    = {MATLAB},
+  title     = {version: 9.14.0 (R2023a)},
   publisher = {The MathWorks Inc.},
   address   = {Natick, Massachusetts, United States},
   url       = {https://www.mathworks.com}
