+++ paper.bib
@@ -47,9 +47,8 @@
 
 @software{MATLAB,
   year      = {2023},
-  author    = {MATLAB},
-  title     = {version: 9.14.0 (R2023a)},
-  publisher = {The MathWorks Inc.},
+  author    = {{The MathWorks Inc.}},
+  title     = {MATLAB version: 9.14.0 (R2023a)},
   address   = {Natick, Massachusetts, United States},
   url       = {https://www.mathworks.com}
 }
