### 2025-05-21 #7748 @SnirMeiri
Hi @rkurchin
I have corrected the paper according to your notes and uploaded the release to Zenodo (and verified the items on the list).
The DOI is 10.5281/zenodo.15479299      https://doi.org/10.5281/zenodo.15479298
The version tag is v1.0.0
Thanks!
