### 2025-05-21 #7748 @SnirMeiri
@rkurchin  you are right, I was confused with the license showing in the pdf of the paper that is created automatically in GitHub. Sorry for that.
I changed it to the same one as in the repository.
