### 2025-05-21 #7748 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
