### 2025-05-23 #7748 @Kevin-Mattheus-Moerman
@SnirMeiri as AEiC for JOSS I will now help to process this submission for acceptance in JOSS.

Yes 👉 https://opensource.org/license/nposl-3-0
